Move _check_type_jsonarg() out of basic.py
Also move JSON realated functions to module_utils.json_utils
