+++ test/units/executor/module_common/test_recursive_finder.py
@@ -53,12 +53,15 @@ MODULE_UTILS_BASIC_IMPORTS = frozenset((('_text',),
                                         ('common', 'sys_info'),
                                         ('common', 'validation'),
                                         ('common', '_utils'),
+                                        ('json_utils',),
                                         ('distro', '__init__'),
                                         ('distro', '_distro'),
                                         ('parsing', '__init__'),
                                         ('parsing', 'convert_bool'),
                                         ('pycompat24',),
                                         ('six', '__init__'),
+                                        ('text', '__init__'),
+                                        ('text', 'formatters'),
                                         ))
 
 MODULE_UTILS_BASIC_FILES = frozenset(('ansible/module_utils/_text.py',
@@ -78,10 +81,13 @@ MODULE_UTILS_BASIC_FILES = frozenset(('ansible/module_utils/_text.py',
                                       'ansible/module_utils/common/_utils.py',
                                       'ansible/module_utils/distro/__init__.py',
                                       'ansible/module_utils/distro/_distro.py',
+                                      'ansible/module_utils/json_utils.py',
                                       'ansible/module_utils/parsing/__init__.py',
                                       'ansible/module_utils/parsing/convert_bool.py',
                                       'ansible/module_utils/pycompat24.py',
                                       'ansible/module_utils/six/__init__.py',
+                                      'ansible/module_utils/text/__init__.py',
+                                      'ansible/module_utils/text/formatters.py',
                                       ))
 
 ONLY_BASIC_IMPORT = frozenset((('basic',),))
